update .net to 8 move entities to separate project delete categories add forbidden categories
