+++ .config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "dotnet-ef": {
-      "version": "7.0.13",
+      "version": "8.0.0",
       "commands": [
         "dotnet-ef"
       ]
