+++ ImageHosting.Persistence/Entities/ForbiddenCategory.cs
@@ -1,0 +1,6 @@
+namespace ImageHosting.Persistence.Entities;
+
+public class ForbiddenCategory
+{
+    public string Name { get; set; } = null!;
+}
