+++ ImageHosting.Persistence/Entities/Image.cs
@@ -1,12 +1,14 @@
-namespace ImageHosting.Storage.Entities;
+namespace ImageHosting.Persistence.Entities;
 
 public class Image 
 {
     public Guid Id { get; set; }
 
     public Guid UserId { get; set; }
+    
     public string ObjectName { get; set; } = null!;
+    public bool Hidden { get; set; }
     public DateTime UploadedAt { get; set; }
-
-    public HashSet<ImageCategory> Categories { get; set; } = null!;
+    
+    public List<string>? Categories { get; set; }
 }
