--- ImageHosting.Storage/Entities/Configuration/ImageCategoryConfiguration.cs
@@ -1,12 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace ImageHosting.Storage.Entities.Configuration;
-
-public class ImageCategoryConfiguration : IEntityTypeConfiguration<ImageCategory>
-{
-    public void Configure(EntityTypeBuilder<ImageCategory> builder)
-    {
-        builder.HasKey(ic => new {ic.ImageId, ic.CategoryId});
-    }
-}
