+++ ImageHosting.sln
@@ -5,9 +5,12 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{2E3ED63C-82E4-4F1D-8053-201A78C2DE9B}"
 	ProjectSection(SolutionItems) = preProject
 		docker-compose.yml = docker-compose.yml
+		.config\dotnet-tools.json = .config\dotnet-tools.json
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImageHosting.Storage.UnitTests", "ImageHosting.Storage.UnitTests\ImageHosting.Storage.UnitTests.csproj", "{56562A88-B4D7-43FD-948C-927A1DA6DDB9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImageHosting.Persistence", "ImageHosting.Persistence\ImageHosting.Persistence.csproj", "{4D2A4C6C-755A-42AA-8366-C8B39564DA37}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,5 +26,9 @@ Global
 		{56562A88-B4D7-43FD-948C-927A1DA6DDB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{56562A88-B4D7-43FD-948C-927A1DA6DDB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{56562A88-B4D7-43FD-948C-927A1DA6DDB9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4D2A4C6C-755A-42AA-8366-C8B39564DA37}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4D2A4C6C-755A-42AA-8366-C8B39564DA37}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4D2A4C6C-755A-42AA-8366-C8B39564DA37}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4D2A4C6C-755A-42AA-8366-C8B39564DA37}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
